requirements.txt
ipython
jinja2
jsonschema
numpy
mahotas
matplotlib
notebook
pandas
pyzmq
scikit-image
scikit-learn
scipy
seaborn
statsmodels
surprise
tornado
plotly
Keras
GitPython
pylint
